requirements.txt
# Core dependencies
python-dotenv>=0.19.0
replicate>=0.15.0
pillow>=9.0.0
pyautogui>=0.9.53
mss>=6.1.0
boto3>=1.24.0
google-generativeai>=0.3.0
openai>=1.10.0

# GUI dependencies
pyqt5>=5.15.6

# Voice dependencies
PyAudio>=0.2.13
pygame>=2.5.0
pyttsx3>=2.90  # Fallback text-to-speech engine

# Optional S3 dependencies
boto3>=1.24.0
botocore>=1.27.0 